Criar arquivos utils para funcoes que nao sao actions ou filters
