+++ delibera.php
@@ -63,32 +63,9 @@ require_once __DIR__.DIRECTORY_SEPARATOR.'delibera_setup.php';
 
 require_once __DIR__.DIRECTORY_SEPARATOR.'delibera_init.php';
 
-/**
-	 *
-	 * Revemos acentos do texto
-	 * @param string $texto
-	 * @return string
-	 */
-function delibera_tiracento($texto)
-{
-	$trocarIsso = 	array('à','á','â','ã','ä','å','ç','è','é','ê','ë','ì','í','î','ï','ñ','ò','ó','ô','õ','ö','ù','ü','ú','ÿ','À','Á','Â','Ã','Ä','Å','Ç','È','É','Ê','Ë','Ì','Í','Î','Ï','Ñ','Ò','Ó','Ô','Õ','Ö','Ù','Ü','Ú','Ÿ',);
-	$porIsso = 		array('a','a','a','a','a','a','c','e','e','e','e','i','i','i','i','n','o','o','o','o','o','u','u','u','y','A','A','A','A','A','A','C','E','E','E','E','I','I','I','I','N','O','O','O','O','O','U','U','U','Y',);
-	$titletext = str_replace($trocarIsso, $porIsso, $texto);
-	return $titletext;
-}
+require_once __DIR__.DIRECTORY_SEPARATOR.'delibera_utils.php';
 
-function delibera_slug_under($label)
-{
-	$slug = delibera_tiracento($label);
-	$slug = str_replace(array("'",'"','.',',',';','!'), '', $slug);
-	$slug = str_replace(array(' ', '-'), '_', $slug);
-	return strtolower($slug);
-}
-
-function is_pauta($post = false)
-{
-	return get_post_type($post) == 'pauta' ? true : false;
-}
+require_once __DIR__.DIRECTORY_SEPARATOR.'delibera_comments.php';
 
 /**
  *
@@ -437,15 +414,6 @@ function delibera_pauta_meta()
 		<input <?php echo $disable_edicao ?> id="prazo_votacao" name="prazo_votacao" class="prazo_votacao widefat hasdatepicker" value="<?php echo $prazo_votacao; ?>"/>
 	</p>
 	<?php
-}
-
-function delibera_tratar_data($data, $int = true, $full = true)
-{
-	$data = trim($data);
-	if(strlen($data) < 8) return false;
-	$data = substr($data, 6, 4).substr($data, 2, 4).substr($data, 0, 2);
-	$data .= $full === true ? ' 23:59:59' : '';
-	return strtotime($data);
 }
 
 /**
@@ -971,157 +939,6 @@ add_filter('wp_insert_post_data', 'delibera_check_post_data', 10, 2);
 
 /**
  *
- *
- * @param string $text String to truncate.
- * @param integer $length Length of returned string, including ellipsis.
- * @param string $ending Ending to be appended to the trimmed string.
- * @param boolean $exact If false, $text will not be cut mid-word
- * @param boolean $considerHtml If true, HTML tags would be handled correctly
- * @return string Trimmed string.
- */
-
-function truncate($text, $length = 100, $ending = '...', $exact = false, $considerHtml = true) {
-	if ($considerHtml) {
-		// if the plain text is shorter than the maximum length, return the whole text
-		if (strlen(preg_replace('/<.*?>/', '', $text)) <= $length) {
-			return $text;
-		}
-		// splits all html-tags to scanable lines
-		preg_match_all('/(<.+?>)?([^<>]*)/s', $text, $lines, PREG_SET_ORDER);
-		$total_length = strlen($ending);
-		$open_tags = array();
-		$truncate = '';
-		foreach ($lines as $line_matchings) {
-			// if there is any html-tag in this line, handle it and add it (uncounted) to the output
-			if (!empty($line_matchings[1])) {
-				// if it's an "empty element" with or without xhtml-conform closing slash
-				if (preg_match('/^<(\s*.+?\/\s*|\s*(img|br|input|hr|area|base|basefont|col|frame|isindex|link|meta|param)(\s.+?)?)>$/is', $line_matchings[1])) {
-					// do nothing
-				// if tag is a closing tag
-				} else if (preg_match('/^<\s*\/([^\s]+?)\s*>$/s', $line_matchings[1], $tag_matchings)) {
-					// delete tag from $open_tags list
-					$pos = array_search($tag_matchings[1], $open_tags);
-					if ($pos !== false) {
-					unset($open_tags[$pos]);
-					}
-				// if tag is an opening tag
-				} else if (preg_match('/^<\s*([^\s>!]+).*?>$/s', $line_matchings[1], $tag_matchings)) {
-					// add tag to the beginning of $open_tags list
-					array_unshift($open_tags, strtolower($tag_matchings[1]));
-				}
-				// add html-tag to $truncate'd text
-				$truncate .= $line_matchings[1];
-			}
-			// calculate the length of the plain text part of the line; handle entities as one character
-			$content_length = strlen(preg_replace('/&[0-9a-z]{2,8};|&#[0-9]{1,7};|[0-9a-f]{1,6};/i', ' ', $line_matchings[2]));
-			if ($total_length+$content_length> $length) {
-				// the number of characters which are left
-				$left = $length - $total_length;
-				$entities_length = 0;
-				// search for html entities
-				if (preg_match_all('/&[0-9a-z]{2,8};|&#[0-9]{1,7};|[0-9a-f]{1,6};/i', $line_matchings[2], $entities, PREG_OFFSET_CAPTURE)) {
-					// calculate the real length of all entities in the legal range
-					foreach ($entities[0] as $entity) {
-						if ($entity[1]+1-$entities_length <= $left) {
-							$left--;
-							$entities_length += strlen($entity[0]);
-						} else {
-							// no more characters left
-							break;
-						}
-					}
-				}
-				$truncate .= substr($line_matchings[2], 0, $left+$entities_length);
-				// maximum lenght is reached, so get off the loop
-				break;
-			} else {
-				$truncate .= $line_matchings[2];
-				$total_length += $content_length;
-			}
-			// if the maximum length is reached, get off the loop
-			if($total_length>= $length) {
-				break;
-			}
-		}
-	} else {
-		if (strlen($text) <= $length) {
-			return $text;
-		} else {
-			$truncate = substr($text, 0, $length - strlen($ending));
-		}
-	}
-	// if the words shouldn't be cut in the middle...
-	if (!$exact) {
-		// ...search the last occurance of a space...
-		$spacepos = strrpos($truncate, ' ');
-		if (isset($spacepos)) {
-			// ...and cut the text in this position
-			$truncate = substr($truncate, 0, $spacepos);
-		}
-	}
-	// add the defined ending to the text
-	$truncate .= $ending;
-	if($considerHtml) {
-		// close all unclosed html-tags
-		foreach ($open_tags as $tag) {
-			$truncate .= '</' . $tag . '>';
-		}
-	}
-	return $truncate;
-}
-
-// Fim Inicialização do plugin
-
-// Menu de configuração
-
-function delibera_config_menu()
-{
-	/*if (function_exists('add_menu_page'))
-		add_menu_page( __('Delibera','delibera'), __('Delibera plugin','delibera'), 'manage_options', 'delibera-config', 'delibera_conf_page');*/
-
-	$base_page = 'delibera-config';
-
-	if (function_exists('add_menu_page'))
-	{
-		add_object_page( __('Delibera','delibera'), __('Delibera','delibera'), 'manage_options', $base_page, array(), WP_PLUGIN_URL."/delibera/images/delibera_icon.png");
-		//add_submenu_page($base_page, __('Pesquisar Contatos','delibera'), __('Pesquisar Contatos','delibera'), 'manage_options', 'delibera-gerenciar', 'delibera_GerenciarContato' );
-		//add_submenu_page($base_page, __('Criar Contato','delibera'), __('Criar Contato','delibera'), 'manage_options', 'delibera-criar', 'delibera_CriarContato' );
-		//add_submenu_page($base_page, __('Importar Contatos','delibera'), __('Importar Contatos','delibera'), 'manage_options', 'delibera-importar', 'delibera_ImportarContato' );
-		add_submenu_page($base_page, __('Configurações do Plugin','delibera'),__('Configurações do Plugin','delibera'), 'manage_options', 'delibera-config', 'delibera_conf_page');
-		do_action('delibera_menu_itens', $base_page);
-	}
-}
-
-/**
- * Create a form table from an array of rows
- */
-function delibera_form_table($rows) {
-	$content = '<table class="form-table">';
-	foreach ($rows as $row) {
-		$content .= '<tr '.(array_key_exists('row-id', $row) ? 'id="'.$row['row-id'].'"' : '' ).' '.(array_key_exists('row-style', $row) ? 'style="'.$row['row-style'].'"' : '' ).' '.(array_key_exists('row-class', $row) ? 'class="'.$row['row-class'].'"' : '' ).' ><th valign="top" scrope="row">';
-
-		if (isset($row['id']) && $row['id'] != '') {
-			$content .= '<label for="'.$row['id'].'">'.$row['label'].'</label>';
-		} else {
-			$content .= $row['label'];
-		}
-
-		if (isset($row['desc']) && $row['desc'] != '') {
-			$content .= '<br/><small>'.$row['desc'].'</small>';
-		}
-
-		$content .= '</th><td valign="top">';
-		$content .= $row['content'];
-		$content .= '</td></tr>';
-	}
-	$content .= '</table>';
-	return $content;
-}
-
-// Funções de conteudo
-
-/**
- *
  * Retorna post do tipo pauta em uma determinada situacao (validacao, discussao, emvotacao ou comresolucao), usando um filtro
  * @param array $filtro
  * @param string $situacao
@@ -1203,262 +1020,6 @@ function delibera_get_resolucoes($filtro = array())
 	return delibera_get_pautas_em($filtro, 'comresolucao');
 }
 
-require_once 'delibera_comments_template.php';
-
-function delibera_get_comments_padrao($args = array(), $file = '/comments.php' )
-{
-	global $delibera_comments_padrao;
-	$delibera_comments_padrao = true;
-	comments_template($file);
-	$delibera_comments_padrao = false;
-}
-
-/**
- * Retorna comentários do Delibera de acordo com o tipo.
- *
- * @param int $post_id
- * @param string|array $tipo um tipo ou um array de tipos
- * @return array
- */
-function delibera_get_comments($post_id, $tipo, $args = array())
-{
-	if (is_string($tipo)) {
-		$tipo = array($tipo);
-	}
-
-	$args = array_merge(array('post_id' => $post_id), $args);
-	$comments = get_comments($args);
-	$ret = array();
-	foreach ($comments as $comment)
-	{
-		$comment_tipo = get_comment_meta($comment->comment_ID, 'delibera_comment_tipo', true);
-		if (in_array($comment_tipo, $tipo)) {
-			$ret[] = $comment;
-		}
-	}
-	return $ret;
-}
-
-require_once __DIR__.DIRECTORY_SEPARATOR.'delibera_comments.php';
-
-
-function delibera_wp_list_comments($args = array(), $comments = null)
-{
-	global $post;
-	global $delibera_comments_padrao;
-
-	if(get_post_type($post) == "pauta")
-	{
-		$situacao = delibera_get_situacao($post->ID);
-
-		if($delibera_comments_padrao === true)
-		{
-			$args['post_id'] = $post->ID;
-			$args['walker'] = new Delibera_Walker_Comment_padrao();
-			$comments = get_comments($args);
-			$ret = array();
-			foreach ($comments as $comment)
-			{
-				$tipo_tmp = get_comment_meta($comment->comment_ID, 'delibera_comment_tipo', true);
-				if(strlen($tipo_tmp) <= 0 || $tipo_tmp === false)
-				{
-					$ret[] = $comment;
-				}
-			}
-			wp_list_comments($args, $ret);
-		}
-		elseif($situacao->slug == 'validacao')
-		{
-			//comment_form();
-			$args['walker'] = new Delibera_Walker_Comment();
-			//$args['callback'] = 'delibera_comments_list';
-			?>
-			<div class="delibera_lista_validacoes">
-			<?php
-			wp_list_comments($args, $comments);
-			?>
-			</div>
-			<?php
-		}
-		elseif($situacao->slug == 'comresolucao')
-		{
-			$args['walker'] = new Delibera_Walker_Comment();
-			wp_list_comments($args, $comments);
-
-			$encaminhamentos = delibera_get_comments_encaminhamentos($post->ID);
-			$discussoes = delibera_get_comments_discussoes($post->ID);
-			?>
-			<div class="delibera_encaminhamentos_inferior">
-    			<?php wp_list_comments($args, $encaminhamentos); ?>
-			</div>
-
-			<div id="comments" class="delibera_opinioes_inferior">
-			    <hr>
-			    <h2 class="comments-title bottom"><?php _e('Histórico da pauta', 'delibera'); ?></h2>
-			    <?php wp_list_comments($args, $discussoes); ?>
-			</div>
-
-			<?php
-		}
-		else
-		{
-			$args['walker'] = new Delibera_Walker_Comment();
-			//$args['callback'] = 'delibera_comments_list';
-			wp_list_comments($args, $comments);
-		}
-	}
-	else
-	{
-		wp_list_comments($args, $comments);
-	}
-}
-
-
-/**
- * Retrieve a list of comments.
- *
- * The comment list can be for the blog as a whole or for an individual post.
- *
- * The list of comment arguments are 'status', 'orderby', 'comment_date_gmt',
- * 'order', 'number', 'offset', and 'post_id'.
- *
- * @since 2.7.0
- * @uses $wpdb
- *
- * @param mixed $args Optional. Array or string of options to override defaults.
- * @return array List of comments.
- */
-function delibera_wp_get_comments( $args = '' ) {
-	$query = new delibera_WP_Comment_Query();
-	return $query->query( $args );
-}
-
-function delibera_get_comments_validacoes($post_id)
-{
-	return delibera_get_comments($post_id, 'validacao');
-}
-
-function delibera_get_comments_discussoes($post_id)
-{
-	return delibera_get_comments($post_id, 'discussao');
-}
-
-function delibera_get_comments_encaminhamentos($post_id)
-{
-	return delibera_get_comments($post_id, 'encaminhamento');
-}
-
-/**
- * Retorna os encaminhamentos dos tipos 'encaminhamento' e
- * 'encaminhamento_selecionado' (aqueles que foram selecionados
- * pelo relator para ir para votação).
- *
- * @param int $post_id
- * @return array
- */
-function delibera_get_comments_all_encaminhamentos($post_id)
-{
-    return delibera_get_comments($post_id, array('encaminhamento', 'encaminhamento_selecionado'));
-}
-
-/**
- * Retorna os encaminhamentos do tipo 'encaminhamento_selecionado'
- * (aqueles que foram selecionados pelo relator para ir para votação).
- *
- * @param int $post_id
- * @return array
- */
-function delibera_get_comments_encaminhamentos_selecionados($post_id)
-{
-    return delibera_get_comments($post_id, 'encaminhamento_selecionado');
-}
-
-
-function delibera_get_comments_votacoes($post_id)
-{
-	return delibera_get_comments($post_id, 'voto');
-}
-
-function delibera_get_comments_resolucoes($post_id)
-{
-	if(has_filter('delibera_get_resolucoes'))
-	{
-		return apply_filters('delibera_get_resolucoes', delibera_get_comments($post_id, 'resolucao'));
-	}
-	return delibera_get_comments($post_id, 'resolucao');
-}
-
-/**
- *
- * Busca comentários com o tipo em tipos
- * @param array $comments lista de comentários a ser filtrada
- * @param array $tipos tipos aceitos
- */
-function delibera_comments_filter_portipo($comments, $tipos)
-{
-	$ret = array();
-
-	foreach ($comments as $comment)
-	{
-		$tipo = get_comment_meta($comment->comment_ID, 'delibera_comment_tipo', true);
-		if(array_search($tipo, $tipos) !== false)
-		{
-			$ret[] = $comment;
-		}
-	}
-	return $ret;
-}
-
-/**
- *
- * Filtro que retorna Comentário filtrados pela a situação da pauta
- * @param array $comments
- * @param int $postID
- * @return array Comentários filtrados
- */
-function delibera_get_comments_filter($comments)
-{
-	global $delibera_comments_padrao;
-
-	if($delibera_comments_padrao === true) return $comments;
-
-	$ret = array();
-
-	if(count($comments) > 0)
-	{
-		if(get_post_type($comments[0]->comment_post_ID) == "pauta")
-		{
-			$situacao = delibera_get_situacao($comments[0]->comment_post_ID);
-			switch ($situacao->slug)
-			{
-				case 'validacao':
-				{
-					$ret = delibera_comments_filter_portipo($comments, array('validacao'));
-				}break;
-				case 'discussao':
-				{
-					$ret = delibera_comments_filter_portipo($comments, array('discussao', 'encaminhamento'));
-				}break;
-				case 'relatoria':
-				{
-					$ret = delibera_comments_filter_portipo($comments, array('discussao', 'encaminhamento'));
-				}break;
-				case 'emvotacao':
-				{
-					$ret = delibera_comments_filter_portipo($comments, array('voto'));
-				}break;
-				case 'comresolucao':
-				{
-					$ret = delibera_comments_filter_portipo($comments, array('resolucao'));
-				}break;
-			}
-			return $ret;
-		}
-	}
-	return $comments;
-}
-
-add_filter('comments_array', 'delibera_get_comments_filter');
 
 /**
  * Retorna um inteiro indicando quantos dias faltam para o fim do prazo
@@ -1557,70 +1118,6 @@ function delibera_post_custom_column($column)
 
 add_action('manage_posts_custom_column',  'delibera_post_custom_column');
 
-function delibera_comment_number($postID, $tipo)
-{
-	switch($tipo)
-	{
-		case 'validacao':
-			return doubleval(get_post_meta($postID, 'delibera_numero_comments_validacoes', true));
-		break;
-		case 'discussao':
-			return doubleval(get_post_meta($postID, 'delibera_numero_comments_discussoes', true));
-		break;
-		case 'encaminhamento':
-			return doubleval(get_post_meta($postID, 'delibera_numero_comments_encaminhamentos', true));
-		break;
-		case 'voto':
-			return doubleval(get_post_meta($postID, 'delibera_numero_comments_votos', true));
-		break;
-		/*case 'resolucao':
-			return doubleval(get_post_meta($postID, 'delibera_numero_comments_resolucoes', true)); TODO Número de resoluções, baseado no mínimo de votos, ou marcação especial
-		break;*/
-		case 'todos':
-			return get_post($postID)->comment_count;
-		break;
-		default:
-			return doubleval(get_post_meta($postID, 'delibera_numero_comments_padroes', true));
-		break;
-	}
-}
-
-function delibera_comment_number_filtro($count, $postID)
-{
-	if (!is_pauta()) {
-		return $count;
-	}
-	$situacao = delibera_get_situacao($postID);
-
-	if (!$situacao) {
-		return;
-	}
-
-	switch($situacao->slug)
-	{
-		case 'validacao':
-			return doubleval(get_post_meta($postID, 'delibera_numero_comments_validacoes', true));
-		break;
-		case 'discussao':
-		case 'comresolucao':
-			return doubleval(
-				get_post_meta($postID, 'delibera_numero_comments_encaminhamentos', true) +
-				get_post_meta($postID, 'delibera_numero_comments_discussoes', true)
-			);
-		break;
-		case 'relatoria':
-			return doubleval(get_post_meta($postID, 'delibera_numero_comments_encaminhamentos', true));
-		break;
-		case 'emvotacao':
-			return doubleval(get_post_meta($postID, 'delibera_numero_comments_votos', true));
-		break;
-		default:
-			return doubleval(get_post_meta($postID, 'delibera_numero_comments_padroes', true));
-		break;
-	}
-}
-
-add_filter('get_comments_number', 'delibera_comment_number_filtro', 10, 2);
 
 function delibera_restrict_listings()
 {
@@ -1977,26 +1474,6 @@ function delibera_set_novo_prazo_discussao_relatoria($postID, $inova_data, $opts
 	delibera_criar_agenda($postID, false, $nova_data, $nova_data_votacao, $nova_relatoria, $nova_eleicao_rel);
 }
 
-
-function delibera_get_plan_config()
-{
-	$plan = 'N';
-
-	if( is_multisite() && get_current_blog_id() != 1 )
-	{
-		switch_to_blog(1);
-		$opt = delibera_get_config();
-		$plan = $opt['plan_restriction'];
-		restore_current_blog();
-	}
-	else
-	{
-		$opt = delibera_get_config();
-		$plan = $opt['plan_restriction'];
-	}
-
-	return $plan;
-}
 
 
 add_action('init', 'delibera_nova_pauta_create_action');
